Correct return type of decode functions
Should be int, not ptrdiff_t

+++ src/aux.h
@@ -83,8 +83,8 @@ state *init_default_stacks(size_t memory_size);
 #define INSTRUCTION_MAX_CHUNKS (((word_bit + INSTRUCTION_CHUNK_BIT - 1) / INSTRUCTION_CHUNK_BIT))
 ptrdiff_t encode_instruction_file(int fd, enum instruction_type type, WORD v);
 ptrdiff_t encode_instruction(state *S, UWORD addr, enum instruction_type type, WORD v);
-ptrdiff_t decode_instruction_file(int fd, WORD *val);
-ptrdiff_t decode_instruction(state *S, UWORD *addr, WORD *val);
+int decode_instruction_file(int fd, WORD *val);
+int decode_instruction(state *S, UWORD *addr, WORD *val);
 
 // Object files
 #define MAGIC_LENGTH 8
